OneD mesh improvements, and SetAllMaterialID
