+++ ChiTech/ChiMesh/OrthoGrids/mesh_orthomacros_02_unpartitioned.cc
@@ -49,7 +49,7 @@ void chi_mesh::CreateUnpartitioned1DOrthoMesh(std::vector<double>& vertices)
 
   //======================================== Create vertices
   umesh->vertices.reserve(zverts.size());
-  for (auto& vertex : vertices)
+  for (auto& vertex : zverts)
     umesh->vertices.push_back(new chi_mesh::Vertex(vertex));
 
   //======================================== Create cells
