+++ ChiTech/ChiMesh/VolumeMesher/chi_volumemesher_01_utils.cc
@@ -23,7 +23,8 @@ extern ChiTimer chi_program_timer;
 //###################################################################
 /**Creates 2D polygon cells for each face of a surface mesh.*/
 void chi_mesh::VolumeMesher::
-  AddContinuumToRegion(chi_mesh::MeshContinuumPtr grid, chi_mesh::Region& region)
+  AddContinuumToRegion(chi_mesh::MeshContinuumPtr& grid,
+                       chi_mesh::Region& region)
 {
   region.volume_mesh_continua.push_back(grid);
 }
@@ -32,143 +33,7 @@ void chi_mesh::VolumeMesher::
 /**Creates 2D polygon cells for each face of a surface mesh.*/
 void chi_mesh::VolumeMesher::
 CreatePolygonCells(chi_mesh::SurfaceMesh *surface_mesh,
-                   chi_mesh::MeshContinuumPtr vol_continuum,
-                   bool delete_surface_mesh_elements,
-                   bool force_local)
-{
-  //============================================= Get current mesh handler
-  chi_mesh::MeshHandler* handler = chi_mesh::GetCurrentHandler();
-
-  //============================================= Copy nodes
-  for (auto& vertex : surface_mesh->vertices)
-    vol_continuum->vertices.push_back(new chi_mesh::Node(vertex));
-
-  //============================================= Delete nodes
-  if (delete_surface_mesh_elements)
-    surface_mesh->vertices = std::vector<chi_mesh::Vertex>(0);
-
-  //============================================= Process faces
-  unsigned int num_cells = 0;
-  for (auto& face : surface_mesh->faces)
-  {
-    auto cell = new chi_mesh::CellPolygon;
-
-    for (int k=0;k<3;k++)
-    {
-      cell->vertex_ids.push_back(face.v_index[k]);
-
-      chi_mesh::CellFace new_face;
-
-      new_face.vertex_ids.push_back(face.e_index[k][0]);
-      new_face.vertex_ids.push_back(face.e_index[k][1]);
-
-
-      chi_mesh::Vertex& v0 = *vol_continuum->vertices[face.e_index[k][0]];
-      chi_mesh::Vertex& v1 = *vol_continuum->vertices[face.e_index[k][1]];
-      new_face.centroid = v0*0.5 + v1*0.5;
-
-      chi_mesh::Vector3 vk = chi_mesh::Vector3(0.0, 0.0, 1.0);
-
-      chi_mesh::Vector3 va = v1 - v0;
-      chi_mesh::Vector3 vn = va.Cross(vk);
-      vn = vn/vn.Norm();
-      new_face.normal = vn;
-
-      if (face.e_index[k][2]>=0)
-      {
-        new_face.neighbor_id = face.e_index[k][2];
-        new_face.has_neighbor = true;
-      }
-
-      cell->faces.push_back(new_face);
-
-      cell->centroid = cell->centroid + surface_mesh->vertices[face.v_index[k]];
-    }
-    cell->centroid = cell->centroid/3;
-
-    //====================================== Compute xy partition id
-    auto xy_partition_indices = GetCellXYPartitionID(cell);
-    cell->partition_id = xy_partition_indices.second*
-                         handler->surface_mesher->partitioning_x +
-                         xy_partition_indices.first;
-
-    if (force_local)
-      cell->partition_id = chi_mpi.location_id;
-
-    cell->global_id = num_cells;
-
-    vol_continuum->cells.push_back(cell); ++num_cells;
-  }
-
-  for (auto face : surface_mesh->poly_faces)
-  {
-    auto cell = new chi_mesh::CellPolygon;
-
-    //====================================== Copy vertices
-    for (auto vid : face->v_indices)
-    {
-      cell->vertex_ids.push_back(vid);
-      cell->centroid = cell->centroid +
-                       *vol_continuum->vertices[vid];
-    }
-    cell->centroid = cell->centroid/cell->vertex_ids.size();
-
-    //====================================== Copy edges
-    for (auto src_side : face->edges)
-    {
-      chi_mesh::CellFace new_face;
-
-      new_face.vertex_ids.push_back(src_side[0]);
-      new_face.vertex_ids.push_back(src_side[1]);
-
-      chi_mesh::Vertex& v0 = *vol_continuum->vertices[src_side[0]];
-      chi_mesh::Vertex& v1 = *vol_continuum->vertices[src_side[1]];
-      new_face.centroid = v0*0.5 + v1*0.5;
-      chi_mesh::Vector3 vk = chi_mesh::Vector3(0.0, 0.0, 1.0);
-
-      chi_mesh::Vector3 va = v1 - v0;
-      chi_mesh::Vector3 vn = va.Cross(vk);
-      vn = vn/vn.Norm();
-      new_face.normal = vn;
-
-      if (src_side[2] >= 0)
-      {
-        new_face.neighbor_id = src_side[2];
-        new_face.has_neighbor = true;
-      }
-      else
-        new_face.neighbor_id = 0;
-
-      cell->faces.push_back(new_face);
-    }
-
-    //====================================== Compute partition id
-    auto xy_partition_indices = GetCellXYPartitionID(cell);
-    cell->partition_id = xy_partition_indices.second*
-                         handler->surface_mesher->partitioning_x +
-                         xy_partition_indices.first;
-
-    if (force_local)
-      cell->partition_id = chi_mpi.location_id;
-
-    cell->global_id = num_cells;
-
-    vol_continuum->cells.push_back(cell); ++num_cells;
-
-    if (delete_surface_mesh_elements)
-      delete face;
-  }
-
-  if (delete_surface_mesh_elements)
-    surface_mesh->poly_faces.clear();
-
-}
-
-//###################################################################
-/**Creates 2D polygon cells for each face of a surface mesh.*/
-void chi_mesh::VolumeMesher::
-CreatePolygonCells(chi_mesh::SurfaceMesh *surface_mesh,
-                   chi_mesh::MeshContinuum* vol_continuum,
+                   chi_mesh::MeshContinuumPtr& vol_continuum,
                    bool delete_surface_mesh_elements,
                    bool force_local)
 {
@@ -303,61 +168,8 @@ CreatePolygonCells(chi_mesh::SurfaceMesh *surface_mesh,
 //###################################################################
 /**Filters non-esential ghosts from the grid.*/
 void chi_mesh::VolumeMesher::
-  GridFilterGhosts(chi_mesh::MeshContinuumPtr in_grid,
-                   chi_mesh::MeshContinuumPtr out_grid)
-{
-  chi_log.Log(LOG_0VERBOSE_1) << "Filtering ghosts.";
-  //======================================== Copy vertices
-  for (auto vertex : in_grid->vertices)
-    out_grid->vertices.push_back(new chi_mesh::Vertex(*vertex));
-
-  //======================================== Copy local cells
-  for (auto& cell : in_grid->local_cells)
-    out_grid->cells.push_back(&cell);
-
-
-
-  //======================================== Copy ghost cells only
-  //                                         if neighbor to current partition
-  auto in_ghost_ids = in_grid->cells.GetGhostGlobalIDs();
-  std::vector<Cell*> cells_to_delete;
-  for (int ghost_id : in_ghost_ids)
-  {
-    auto& ref_ghost_cell = in_grid->cells[ghost_id];
-
-    bool is_neighbor_to_this_loc = false;
-    for (auto& face : ref_ghost_cell.faces)
-    {
-      if (not face.has_neighbor) continue;
-
-      auto adj_cell = in_grid->cells[face.neighbor_id];
-      if (adj_cell.partition_id == chi_mpi.location_id)
-      {
-        is_neighbor_to_this_loc = true;
-        break;
-      }
-    }//for face
-
-    if (is_neighbor_to_this_loc)
-      out_grid->cells.push_back(&ref_ghost_cell);
-    else
-      cells_to_delete.push_back(&ref_ghost_cell);
-  }//for cell
-
-
-  //======================================== Deleting non-essential ghosts
-  for (auto cell : cells_to_delete)
-    delete cell;
-
-
-  chi_log.Log(LOG_0VERBOSE_1) << "Done filtering ghosts.";
-}
-
-//###################################################################
-/**Filters non-esential ghosts from the grid.*/
-void chi_mesh::VolumeMesher::
-GridFilterGhosts(chi_mesh::MeshContinuum* in_grid,
-                 chi_mesh::MeshContinuumPtr out_grid)
+  GridFilterGhosts(chi_mesh::MeshContinuumPtr& in_grid,
+                   chi_mesh::MeshContinuumPtr& out_grid)
 {
   chi_log.Log(LOG_0VERBOSE_1) << "Filtering ghosts.";
   //======================================== Copy vertices
@@ -417,7 +229,6 @@ std::pair<int,int> chi_mesh::VolumeMesher::
  GetCellXYPartitionID(chi_mesh::Cell *cell)
 {
   std::pair<int,int> ij_id(0,0);
-  bool found_partition = false;
 
   if (chi_mpi.process_count == 1){return ij_id;}
 
@@ -428,11 +239,11 @@ std::pair<int,int> chi_mesh::VolumeMesher::
   if  (typeid(*surf_mesher) == typeid(chi_mesh::SurfaceMesherPredefined))
   {
     //====================================== Sanity check on partitioning
-    int num_x_subsets = surf_mesher->xcuts.size()+1;
-    int num_y_subsets = surf_mesher->ycuts.size()+1;
+    size_t num_x_subsets = surf_mesher->xcuts.size()+1;
+    size_t num_y_subsets = surf_mesher->ycuts.size()+1;
 
-    int x_remainder = num_x_subsets%surf_mesher->partitioning_x;
-    int y_remainder = num_y_subsets%surf_mesher->partitioning_y;
+    size_t x_remainder = num_x_subsets%surf_mesher->partitioning_x;
+    size_t y_remainder = num_y_subsets%surf_mesher->partitioning_y;
 
     if (x_remainder != 0)
     {
@@ -450,8 +261,8 @@ std::pair<int,int> chi_mesh::VolumeMesher::
       exit(EXIT_FAILURE);
     }
 
-    int subsets_per_partitionx = num_x_subsets/surf_mesher->partitioning_x;
-    int subsets_per_partitiony = num_y_subsets/surf_mesher->partitioning_y;
+    size_t subsets_per_partitionx = num_x_subsets/surf_mesher->partitioning_x;
+    size_t subsets_per_partitiony = num_y_subsets/surf_mesher->partitioning_y;
 
 //    chi_log.Log(LOG_0ERROR) << num_x_subsets;
 //    chi_log.Log(LOG_0ERROR) << num_y_subsets;
@@ -462,9 +273,9 @@ std::pair<int,int> chi_mesh::VolumeMesher::
     //====================================== Determine x-partition
     int x=-1;
     int xcount=-1;
-    for (int i =  subsets_per_partitionx-1;
-             i <  surf_mesher->xcuts.size();
-             i += subsets_per_partitionx)
+    for (size_t i =  subsets_per_partitionx-1;
+                i <  surf_mesher->xcuts.size();
+                i += subsets_per_partitionx)
     {
       xcount++;
       if (cell->centroid.x <= surf_mesher->xcuts[i])
@@ -481,9 +292,9 @@ std::pair<int,int> chi_mesh::VolumeMesher::
     //====================================== Determine y-partition
     int y=-1;
     int ycount=-1;
-    for (int i =  subsets_per_partitiony-1;
-         i <  surf_mesher->ycuts.size();
-         i += subsets_per_partitiony)
+    for (size_t i =  subsets_per_partitiony-1;
+                i <  surf_mesher->ycuts.size();
+                i += subsets_per_partitiony)
     {
       ycount++;
       if (cell->centroid.y <= surf_mesher->ycuts[i])
@@ -511,7 +322,7 @@ std::pair<int,int> chi_mesh::VolumeMesher::
  * Cell xy_partition ids are obtained from
  * the surface mesher. z id is obtained from the volume mesher.*/
 std::tuple<int,int,int> chi_mesh::VolumeMesher::
-  GetCellXYZPartitionID(chi_mesh::Cell *cell)
+  GetCellXYZPartitionID(chi_mesh::Cell *cell) const
 {
   std::tuple<int,int,int> ijk_id(0,0,0);
   bool found_partition = false;
@@ -520,7 +331,6 @@ std::tuple<int,int,int> chi_mesh::VolumeMesher::
 
   //================================================== Get ij indices
   std::pair<int,int> ij_id = GetCellXYPartitionID(cell);
-
 
   //================================================== Get the current handler
   chi_mesh::MeshHandler*  mesh_handler = chi_mesh::GetCurrentHandler();
@@ -566,7 +376,7 @@ std::tuple<int,int,int> chi_mesh::VolumeMesher::
     //====================================== Create virtual cuts
     if (vol_mesher->zcuts.empty())
     {
-      int num_sub_layers = extruder->vertex_layers.size()-1;
+      size_t num_sub_layers = extruder->vertex_layers.size()-1;
 
       if ((num_sub_layers%vol_mesher->options.partition_z) != 0)
       {
@@ -602,10 +412,9 @@ std::tuple<int,int,int> chi_mesh::VolumeMesher::
 
     //====================================== Scan cuts for location
     double zmin = -1.0e-16;
-    double zmax =  1.0e-16;
     for (int k=0; k<(vol_mesher->zcuts.size()); k++)
     {
-      zmax =  vol_mesher->zcuts[k];
+      double zmax =  vol_mesher->zcuts[k];
 
       double z = cell->centroid.z;
 
@@ -639,32 +448,35 @@ std::tuple<int,int,int> chi_mesh::VolumeMesher::
     }
 
     //====================================== Scan cuts for location
-    double zmin = -1.0e16;
-    double zmax =  1.0e16;
-    vol_mesher->zcuts.push_back(zmax);
-    for (int k=0; k<(vol_mesher->zcuts.size()); k++)
+    if (not found_partition)
     {
-      zmax =  vol_mesher->zcuts[k];
-
-      double z = cell->centroid.z;
-
-      if (chi_log.GetVerbosity()==LOG_0VERBOSE_2)
+      double zmin = -1.0e16;
+      double zmax =  1.0e16;
+      vol_mesher->zcuts.push_back(zmax);
+      for (int k=0; k<(vol_mesher->zcuts.size()); k++)
       {
-        printf("zmax = %g, zmin = %g, cell_z = %g\n",zmax,zmin,z);
-      }
+        zmax =  vol_mesher->zcuts[k];
+
+        double z = cell->centroid.z;
+
+        if (chi_log.GetVerbosity()==LOG_0VERBOSE_2)
+        {
+          printf("zmax = %g, zmin = %g, cell_z = %g\n",zmax,zmin,z);
+        }
 
 
-      if ((z > zmin) && (z < zmax))
-      {
-        std::get<0>(ijk_id) = ij_id.first;
-        std::get<1>(ijk_id) = ij_id.second;
-        std::get<2>(ijk_id) = k;
+        if ((z > zmin) && (z < zmax))
+        {
+          std::get<0>(ijk_id) = ij_id.first;
+          std::get<1>(ijk_id) = ij_id.second;
+          std::get<2>(ijk_id) = k;
 
-        found_partition = true;
-        break;
-      }
-      zmin = zmax;
-    }
+          found_partition = true;
+          break;
+        }
+        zmin = zmax;
+      }//for k
+    }//if not found_partition
   }//if typeid
   else if (typeid(*vol_mesher) == typeid(chi_mesh::VolumeMesherPredefined2D))
   {
@@ -685,35 +497,6 @@ std::tuple<int,int,int> chi_mesh::VolumeMesher::
 
   return ijk_id;
 }
-
-
-
-
-//###################################################################
-/**Get a list of boundary cells.*/
-void chi_mesh::VolumeMesher::
- GetBoundaryCells(chi_mesh::MeshContinuumPtr vol_continuum)
-{
-  for (const auto& cell : vol_continuum->local_cells)
-  {
-
-    for (int f=0; f<cell.faces.size(); f++)
-    {
-      if (not cell.faces[f].has_neighbor)
-      {
-        vol_continuum->boundary_cell_indices.push_back(cell.global_id);
-        break;
-      }
-    }
-  }//for local item_id
-  printf("Number of boundary item_id: %lu\n",vol_continuum->boundary_cell_indices.size());
-
-  vol_continuum->ExportCellsToPython(
-    "BoundaryCells.py",true,
-    &vol_continuum->boundary_cell_indices);
-}
-
-
 
 
 //###################################################################
@@ -749,7 +532,7 @@ void chi_mesh::VolumeMesher::
 //###################################################################
 /**Sets material id's using a logical volume.*/
 void chi_mesh::VolumeMesher::
- SetBndryIDFromLogical(chi_mesh::LogicalVolume *log_vol,bool sense, int bndry_id)
+  SetBndryIDFromLogical(chi_mesh::LogicalVolume *log_vol,bool sense, int bndry_id)
 {
   chi_log.Log(LOG_0)
     << chi_program_timer.GetTimeString()
@@ -778,4 +561,29 @@ void chi_mesh::VolumeMesher::
     << chi_program_timer.GetTimeString()
     << " Done setting boundary id from logical volume. "
     << "Number of faces modified = " << num_faces_modified << ".";
+}
+
+//###################################################################
+/**Sets material id's for all cells to the specified material id.*/
+void chi_mesh::VolumeMesher::
+  SetMatIDToAll(int mat_id)
+{
+  chi_log.Log(LOG_0)
+    << chi_program_timer.GetTimeString()
+    << " Setting material id " << mat_id << "to all cells.";
+
+  //============================================= Get current mesh handler
+  auto handler = chi_mesh::GetCurrentHandler();
+
+  //============================================= Get back mesh
+  auto cur_region = handler->region_stack.back();
+  auto vol_cont = cur_region->GetGrid();
+
+  for (auto& cell : vol_cont->local_cells)
+    cell.material_id = mat_id;
+
+  MPI_Barrier(MPI_COMM_WORLD);
+  chi_log.Log(LOG_0)
+    << chi_program_timer.GetTimeString()
+    << " Done setting material id " << mat_id << " to all cells";
 }
