+++ ChiTest/Diffusion1D.lua
@@ -3,28 +3,43 @@ print("############################################### LuaTest")
 
 
 
+-- --############################################### Setup mesh
+-- chiMeshHandlerCreate()
+--
+-- mesh={}
+-- N=100
+-- L=2.0
+-- xmin = -1.0
+-- dx = L/N
+-- for i=1,(N+1) do
+--     k=i-1
+--     mesh[i] = xmin + k*dx
+-- end
+-- line_mesh = chiLineMeshCreateFromArray(mesh)
+--
+--
+-- region1 = chiRegionCreate()
+-- chiRegionAddLineBoundary(region1,line_mesh);
+--
+--
+-- --############################################### Create meshers
+-- chiSurfaceMesherCreate(SURFACEMESHER_PREDEFINED);
+-- chiVolumeMesherCreate(VOLUMEMESHER_LINEMESH1D);
+--
 --############################################### Setup mesh
 chiMeshHandlerCreate()
 
 mesh={}
 N=100
 L=2.0
-xmin = -1.0
+xmin = 0.0
 dx = L/N
 for i=1,(N+1) do
     k=i-1
     mesh[i] = xmin + k*dx
 end
-line_mesh = chiLineMeshCreateFromArray(mesh)
-
-
-region1 = chiRegionCreate()
-chiRegionAddLineBoundary(region1,line_mesh);
-
-
---############################################### Create meshers
-chiSurfaceMesherCreate(SURFACEMESHER_PREDEFINED);
-chiVolumeMesherCreate(VOLUMEMESHER_LINEMESH1D);
+chiMeshCreateUnpartitioned1DOrthoMesh(mesh)
+chiVolumeMesherExecute();
 
 
 --############################################### Execute meshing
